requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.7.0
websockets==12.0
httpx==0.27.0
grpcio==1.62.0
grpcio-tools==1.62.0
diart
pyannote.audio>=4.0.0
sounddevice
omegaconf
